task 15. 1 certification.
One of the basic sources of information on the spread of infection is a dataset maintained by Johns Hopkins University (https://github.com/CSSEGISandData/COVID-19 ). Using the methods studied in previous classes, I conducted a clustering of countries by morbidity and mortality over the past day. Visualized the results. And I realized where Russia was